Time: 0 ms (100%), Space: 11.6 MB (61.65%) - LeetHub

+++ 1967-number-of-strings-that-appear-as-substrings-in-word/1967-number-of-strings-that-appear-as-substrings-in-word.cpp
@@ -3,7 +3,7 @@ class Solution
 public:
     int numOfStrings(vector <string> &patterns, string word)
     {
-        int ans = 0;
+        short int ans = 0;
         for (int i = 0; i < patterns.size(); i++)
         {
             if (word.contains(patterns[i]))
